+++ benchmarks/tst3d_s_o2_thermal_plasma.py
@@ -130,5 +130,19 @@ for direction in ["forward", "backward", "both", "canceling"]:
 	    every = 40,
 	    time_average = 30
 	)
+	DiagScreen(
+	    shape = "cylinder",
+	    point = [Lx/2., Ly/3., Lz/3.],
+	    vector = [Ly*0.4, 0.1, 0.1],
+	    direction = direction,
+	    deposited_quantity = "weight",
+	    species = ["electron"],
+	    axes = [
+	    	["a", -Lx/2., Lx/2., 10],
+	    	["phi", -math.pi, math.pi, 10],
+	    	],
+	    every = 40,
+	    time_average = 30
+	)